myfav_lite_server/requirements.txt
Django==3.0.4
django-filter==2.2.0
django-taggit==1.2.0
djangorestframework==3.11.0
Markdown==3.2.1
